Detailed commit message on previous changes and I2C implementation
The Slave-Pi is set as the master and the Arduino as the slave with an
address of 0x04. Circuit design is as shown in the "Circuit" section.

Servo controlling algorithms now implemented as interrupts - when the
Slave-Pi writes to the Arduino.

TODO: Map letters to a number between 0 and 53. Rewrite algorithm from
using arrays to using the map dictionary