Redesigned stl files
Bump module failed to print at UOttawa 3d printer. Redesigned it so
there is a 100% chance that it can print. Also redesigned the pusher
component to look like a braille bump at the top and inceased the
diameter of the hole in the pinion gear to make sure the servo
connection can be made.